:construction: Add support for SoundCloud downloads
SoundCloud was incorporated as a supported platform for video downloads. Changes include adding SoundCloud to the list of supported platforms in the PvServiceValidation file. New configuration files for SoundCloud were added to enable this feature, including SoundCloudConfigs.kt, SoundCloudConditions.kt, and SoundCloudConstants.kt. Lastly, the application.yml file was updated to reflect these changes and allow for SoundCloud's configuration.
However, a proper downloading implementation is not done yet