:memo: Updated application.yml and added README.md for vvd-extractor
Improved the comments in the application.yml file for a clearer understanding of each configuration field, making it easy to configure the application for new users.
A README.md file was added to provide detailed information for the Extractor module: its functionalities, prerequisites, and instructions for setting up and running the module. The README also gives an explicit description of the configurations in the application.yml file. This update enhances comprehension and user experience.